refactor(lotto): Domain, View 분리
