+++ src/main/java/lotto/domain/NumberChecker.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 public class NumberChecker {
     private LottoTicket winnerNumbers;
